+++ internal/pkg/instrumentation/bpf/go.opentelemetry.io/auto/sdk/probe.go
@@ -8,10 +8,10 @@ import (
 	"encoding/binary"
 	"log/slog"
 
-	"github.com/cilium/ebpf/perf"
 	"go.opentelemetry.io/collector/pdata/ptrace"
 	"go.opentelemetry.io/otel/trace"
 
+	"go.opentelemetry.io/auto/internal/pkg/instrumentation/perf"
 	"go.opentelemetry.io/auto/internal/pkg/instrumentation/probe"
 	"go.opentelemetry.io/auto/internal/pkg/structfield"
 )